+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - **`[knf]`** Deprecated methods have been removed
 - **`[knf/validators]`** Deprecated validators have been removed
 - **`[log]`** Deprecated variables have been removed
+- **`[sliceutil]`** Deprecated methods have been removed
 
 ### [13.28.2](https://kaos.sh/ek/13.28.2)